Copied methods from exercise IntList into own IntList
