+++ Lists/IntList.java
@@ -1,14 +1,5 @@
-import java.util.Formatter;
-
-/**
- * A naked recursive list of integers, similar to what we saw in lecture 3, but
- * with a large number of additional methods.
- *
- * @author P. N. Hilfinger, with some modifications by Josh Hug and melaniecebula
- *         [Do not modify this file.]
- */
 public class IntList {
-    /**
+  /**
      * First element of list.
      */
     public int first;
@@ -16,6 +7,8 @@ public class IntList {
      * Remaining elements of list.
      */
     public IntList rest;
+
+
 
     /**
      * A List with first FIRST0 and rest REST0.
@@ -33,7 +26,73 @@ public class IntList {
         this(0, null);
     }
 
-    /**
+	/** Return the size of the list using... recursion! */
+	public int size() {
+		if(this.rest == null){
+			return 1;
+		}
+		return 1 + this.rest.size();
+	}
+
+	/** Return the size of the list using no recursion! */
+	public int iterativeSize() {
+		IntList p = this;
+		int size = 0;
+		while(p!= null){
+			size += 1;
+			p = p.rest;
+		}
+		return size;
+	}
+
+	/** Returns the ith value in this list.*/
+	public int get(int i) {
+		if(i==1){
+			return first;
+		}
+		return this.rest.get(i-1);
+	}
+	
+	/** Returns an IntList identical to L, but with
+      * each element incremented by x. L is not allowed
+      * to change. */
+    public static IntList incrList(IntList L, int x) {
+		int listSize = L.size();
+		IntList M = new IntList(L.get(listSize)+x,null);
+		
+		for(int i=1; i<listSize; i+=1){
+			// code here
+			M = new IntList(L.get(listSize-i)+x,M);
+		}
+
+		return M;        
+    }
+
+    /** Returns an IntList identical to L, but with
+      * each element incremented by x. Not allowed to use
+      * the 'new' keyword. */
+    public static IntList dincrList(IntList L, int x) {
+		int listSize = L.size();
+		
+		while(listSize > 0){
+			L.first = L.first + x;
+			L = L.rest;
+			listSize -= 1;
+		}
+
+        return L;
+    }
+
+	/** Prints an IntList from start to finish */
+	public void printList(){
+		int listSize = this.size();
+		
+		for(int i=1; i<=listSize; i+=1){
+			System.out.println(i + " --> " + this.get(i));
+		}
+	}
+	
+	   /**
      * Returns a list equal to L with all elements squared. Destructive.
      */
     public static void dSquareList(IntList L) {
@@ -60,7 +119,7 @@ public class IntList {
             ptr = ptr.rest;
         }
         return res;
-    }
+    } 
 
     /**
      * Returns a list equal to L with all elements squared. Non-destructive.
@@ -95,139 +154,31 @@ public class IntList {
     }
 
 
+	public static void main(String[] args) {
+		IntList L = new IntList(15, null);
+		L = new IntList(10, L);
+		L = new IntList(5, L);
+		L = new IntList(0, L);
+		L = new IntList(-10, L);
+		L = new IntList(-5, L);
+		L = new IntList(-15, L);
+		L = new IntList(35, L);
+		int i = 3;
 
+		// incrList(L,i);
+		// System.out.println("Increment List method called");
+		// L.printList();
 
-
-
-
-
-
-
-
-
-
-
-
-
-    /**
-     * DO NOT MODIFY ANYTHING BELOW THIS LINE! Many of the concepts below here
-     * will be introduced later in the course or feature some form of advanced
-     * trickery which we implemented to help make your life a little easier for
-     * the lab.
-     */
-
-    @Override
-    public int hashCode() {
-        return first;
-    }
-
-    /**
-     * Returns a new IntList containing the ints in ARGS. You are not
-     * expected to read or understand this method.
-     */
-    public static IntList of(Integer... args) {
-        IntList result, p;
-
-        if (args.length > 0) {
-            result = new IntList(args[0], null);
-        } else {
-            return null;
-        }
-
-        int k;
-        for (k = 1, p = result; k < args.length; k += 1, p = p.rest) {
-            p.rest = new IntList(args[k], null);
-        }
-        return result;
-    }
-
-    /**
-     * Returns true iff X is an IntList containing the same sequence of ints
-     * as THIS. Cannot handle IntLists with cycles. You are not expected to
-     * read or understand this method.
-     */
-    public boolean equals(Object x) {
-        if (!(x instanceof IntList)) {
-            return false;
-        }
-        IntList L = (IntList) x;
-        IntList p;
-
-        for (p = this; p != null && L != null; p = p.rest, L = L.rest) {
-            if (p.first != L.first) {
-                return false;
-            }
-        }
-        if (p != null || L != null) {
-            return false;
-        }
-        return true;
-    }
-
-    /**
-     * If a cycle exists in the IntList, this method
-     * returns an integer equal to the item number of the location where the
-     * cycle is detected.
-     * <p>
-     * If there is no cycle, the number 0 is returned instead. This is a
-     * utility method for lab2. You are not expected to read, understand, or
-     * even use this method. The point of this method is so that if you convert
-     * an IntList into a String and that IntList has a loop, your computer
-     * doesn't get stuck in an infinite loop.
-     */
-
-    private int detectCycles(IntList A) {
-        IntList tortoise = A;
-        IntList hare = A;
-
-        if (A == null) {
-            return 0;
-        }
-
-        int cnt = 0;
-
-
-        while (true) {
-            cnt++;
-            if (hare.rest != null) {
-                hare = hare.rest.rest;
-            } else {
-                return 0;
-            }
-
-            tortoise = tortoise.rest;
-
-            if (tortoise == null || hare == null) {
-                return 0;
-            }
-
-            if (hare == tortoise) {
-                return cnt;
-            }
-        }
-    }
-
-    @Override
-    /** Outputs the IntList as a String. You are not expected to read
-     * or understand this method. */
-    public String toString() {
-        Formatter out = new Formatter();
-        String sep;
-        sep = "(";
-        int cycleLocation = detectCycles(this);
-        int cnt = 0;
-
-        for (IntList p = this; p != null; p = p.rest) {
-            out.format("%s%d", sep, p.first);
-            sep = ", ";
-
-            cnt++;
-            if ((cnt > cycleLocation) && (cycleLocation > 0)) {
-                out.format("... (cycle exists) ...");
-                break;
-            }
-        }
-        out.format(")");
-        return out.toString();
-    }
-}
+		// System.out.println(L.size());
+		// System.out.println(L.iterativeSize());
+		// L.printList();
+		// System.out.println("The " + i + "'th value in this list is: " + L.get(i));
+		
+        // Test your answers by uncommenting. Or copy and paste the
+        // code for incrList and dincrList into IntList.java and
+        // run it in the visualizer.
+        // System.out.println(L.get(1));
+        // System.out.println(incrList(L, 3));
+        System.out.println(dincrList(L, 3));
+	}
+}
